WIP compatibility changes for HIRAX Code
- Removed tsys key as it wasn't work with the HIRAX subclasses
- Added some integer casting to frequency indexing steps
- Force the use of the class defined lmax in the BTM generation

+++ drift/pipeline/timestream.py
@@ -805,11 +805,8 @@ def simulate(m, outdir, maps=[], ndays=None, resolution=0, seed=None, **kwargs):
     if ndays > 0:
 
         # Fetch the noise powerspectrum
-        noise_ps = tel.noisepower(
-            np.arange(tel.npairs)[:, np.newaxis],
-            np.array(local_freq)[np.newaxis, :],
-            ndays=ndays,
-        ).reshape(tel.npairs, lfreq)[:, :, np.newaxis]
+        noise_ps = tel.noisepower(np.arange(tel.npairs)[:, np.newaxis], np.array(local_freq, dtype=int)[np.newaxis, :], ndays=ndays).reshape(tel.npairs, lfreq)[:, :, np.newaxis]
+
 
         # Seed random number generator to give consistent noise
         if seed is not None:
